Enum flowrlib::debug_command::DebugCommand

pub enum DebugCommand {

    Ack,
    Breakpoint(Option<BreakpointSpec>),
    Continue,
    DebugClientStarting,
    Delete(Option<BreakpointSpec>),
    Error(String),
    ExitDebugger,
    FunctionList,
    InspectFunction(usize),
    Inspect,
    InspectInput(usize, usize),
    InspectOutput(usize, String),
    InspectBlock(Option<usize>, Option<usize>),
    Invalid,
    List,
    Modify(Option<Vec<String>>),
    RunReset,
    Step(Option<usize>),
    Validate,

}

A Command sent by the debug_client to the debugger

Variants

Ack

Acknowledge event processed correctly

Breakpoint(Option<BreakpointSpec>)

Set a breakpoint - with an optional parameter

Continue

continue execution of the flow

DebugClientStarting

Debug client is starting

Delete(Option<BreakpointSpec>)

delete an existing breakpoint - with an optional parameter

Error(String)

An error on the client side

ExitDebugger

exit the debugger and runtime

FunctionList

List of all functions

InspectFunction(usize)

inspect a function

Inspect

Inspect overall state

InspectInput(usize, usize)

Inspect an Input (function_id, input_number)

InspectOutput(usize, String)

Inspect an Output (function_id, sub-path)

InspectBlock(Option<usize>, Option<usize>)

Inspect a Block (optional source function_id, optional destination function_id)

Invalid

Invalid - used when deserialization goes wrong

List

list existing breakpoints

Modify(Option<Vec<String>>)

modify a debugger or runtime state value e.g. jobs=1 to set parallel jobs to 1

RunReset

reset flow execution back to the initial state, or run the flow from the start

Step(Option<usize>)

step forward in flow execution by executing one (default) or more Jobs

Validate

validate the current state