Module flowrlib::services

Provides well-known service names used across multiple binary crates

Constants

CLIENT_SERVER_DISCOVERY_PORT

This is a “well known” port for announcing and discovering the client/server services

CONTEXT_JOB_SERVICE_NAME

CONTEXT_JOB_SERVICE_NAME can be used to discover the queue serving context jobs for execution

CONTROL_SERVICE_NAME

CONTROL_SERVICE_NAME is a control PUB/SUB socket used to control executors that are listening on the JOB_SERVICE and sending results back via the RESULTS_SERVICE

DEBUG_SERVICE_NAME

DEBUG_SERVICE_NAME can be used to discover the debug service by name

JOB_QUEUES_DISCOVERY_PORT

This is a “well known” port for announcing and discovering the job queues

JOB_SERVICE_NAME

JOB_SERVICE_NAME can be used to discover the queue serving jobs for execution

RESULTS_JOB_SERVICE_NAME

RESULTS_JOB_SERVICE_NAME can be used to discover the queue where to send job results

RUNTIME_SERVICE_NAME

RUNTIME_SERVICE_NAME can be used to discover the runtime service by name

WAIT

WAIT for a message to arrive when performing a receive()

Statics

DONT_WAIT

Do NOT WAIT for a message to arrive when performing a receive()

Functions

discover_service

Try to discover a server offering a particular service by name

enable_service_discovery

Start a background thread that sends out beacons for service discovery by a client every second