Crate flowrlib

Runtime library for flow execution. This will be linked with other code to produce a or runner, such as flowr command line runner.

It is responsible for reading a flow definition in a Manifest file, loading the required libraries from LibraryManifest files and then coordinating the execution by dispatching Jobs to be executed by Function Implementations, providing them the Inputs required to run and gathering the Outputs produced and passing those Outputs to other connected Functions in the network of Functions.

Modules

block

module providing Block struct from runtime that is required for debugging and tracing

coordinator

coordinator is the module that coordinates the execution of flows submitted to it

debug_command

debug_command provides the DebugCommand enum for commands from debug client to debug server

dispatcher

Dispatcher module takes care of dispatching jobs for execution and gathering results

errors

We’ll put our errors in an errors module, and other modules in this crate will use errors::*; to get access to everything error_chain creates.

executor

Executor module receives jobs for execution, executes them and returns results

info

info offers methods to get information about this library

job

module providing Job struct from runtime that is required for debugging and tracing

protocols

Trait for a set of methods a server using the library must supply

run_state

Structure that defines/tracks the current runtime state

services

Service names common across multiple binary crates