Crate flowrlib

Runtime library for flow execution. This will be linked with other code to produce a or runner, such as flowr command line runner.

It is responsible for reading a flow definition in a Manifest file, loading the required libraries from LibraryManifest files and then coordinating the execution by dispatching Jobs to be executed by Function Implementations, providing them the Inputs required to run and gathering the Outputs produced and passing those Outputs to other connected Functions in the network of Functions.

Modules

client_server

coordinator
coordinator is the module that coordinates the execution of flows submitted to it

debug
'debug' defines structs passed between the Server and the Client regarding debug events and client responses to them

info
info offers methods to get information about this library

loader
loader is responsible for loading a flow from it's manifest and loading libraries it uses

metrics

runtime
'runtime' defines structs passed between the Server and the Client regarding runtime events and client responses to them

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.